+++ feature/login/impl/src/main/java/com/teamyg/parfait/feature/login/impl/navigation/EntryBuilder.kt
@@ -1,4 +1,4 @@
-package com.teamyg.parfait.feature.login.impl
+package com.teamyg.parfait.feature.login.impl.navigation
 
 import androidx.compose.foundation.layout.fillMaxSize
 import androidx.compose.foundation.layout.padding
@@ -8,6 +8,7 @@ import androidx.navigation3.runtime.EntryProviderScope
 import androidx.navigation3.runtime.NavKey
 import com.teamyg.parfait.feature.login.api.NavKeyLogin
 import com.teamyg.parfait.core.navigation.Navigator
+import com.teamyg.parfait.feature.login.impl.route.LoginRoute
 
 fun EntryProviderScope<NavKey>.featureLoginEntryBuilder(navigator: Navigator) {
     entry<NavKeyLogin> {
