+++ feature/login/impl/src/main/java/com/teamyg/parfait/feature/login/impl/navigation/NavigationModule.kt
@@ -1,4 +1,4 @@
-package com.teamyg.parfait.feature.login.impl
+package com.teamyg.parfait.feature.login.impl.navigation
 
 import androidx.navigation3.runtime.EntryProviderScope
 import androidx.navigation3.runtime.NavKey
