+++ feature/login/impl/src/main/java/com/teamyg/parfait/feature/login/impl/screen/LoginScreen.kt
@@ -1,4 +1,4 @@
-package com.teamyg.parfait.feature.login.impl
+package com.teamyg.parfait.feature.login.impl.screen
 
 import androidx.compose.foundation.clickable
 import androidx.compose.foundation.layout.Arrangement
