+++ feature/login/impl/src/main/java/com/teamyg/parfait/feature/login/impl/viewmodel/LoginViewModel.kt
@@ -1,4 +1,4 @@
-package com.teamyg.parfait.feature.login.impl
+package com.teamyg.parfait.feature.login.impl.viewmodel
 
 import android.util.Log
 import androidx.lifecycle.viewModelScope
